refactor(bb-lambda-compute): Consolidate observability into observability.ts
Rename widgets.ts to observability.ts and fold the X-Ray tracing activation
(previously inline in LambdaCompute.applyTracing) into it, so all of the
compute's observability — tracing infra plus the dashboard health/logs/traces
widget builders — lives in one module. The class body now delegates:
applyTracing() calls applyXRayTracing(), keeping index.cdk.ts focused on the
compute's core function + gateway wiring.

No behavior or public API change (observability.ts is package-internal).
